+++ sopa/annotation/tangram/run.py
@@ -10,13 +10,6 @@ import torch
 from anndata import AnnData
 from spatialdata import SpatialData
 
-try:
-    import tangram as tg
-except ImportError:
-    raise ImportError(
-        "To use tangram, you need its corresponding sopa extra: `pip install 'sopa[tangram]'` (normal mode) or `pip install -e '.[tangram]'` (if using snakemake)"
-    )
-
 from sopa._constants import SopaKeys
 
 log = logging.getLogger(__name__)
@@ -216,6 +209,13 @@ class MultiLevelAnnotation:
         ]
 
     def run_group(self, level: int = 0, indices_sp=None, indices_sc=None):
+        try:
+            import tangram as tg
+        except ImportError:
+            raise ImportError(
+                "To use tangram, you need its corresponding sopa extra: `pip install 'sopa[tangram]'` (normal mode) or `pip install -e '.[tangram]'` (if using snakemake)"
+            )
+
         if indices_sp is not None and len(indices_sp) == 0:
             log.warn("No cell annotated in the upper level...")
             return