+++ sopa/io/__init__.py
@@ -1,4 +1,17 @@
-from .explorer import write, align, add_explorer_selection
+from .explorer import (
+    write,
+    align,
+    add_explorer_selection,
+    write_image,
+    write_transcripts,
+    write_cell_categories,
+    write_gene_counts,
+    write_polygons,
+    write_metadata,
+    str_cell_id,
+    int_cell_id,
+    save_column_csv,
+)
 from .standardize import write_standardized
 from .reader.cosmx import cosmx
 from .reader.merscope import merscope
@@ -8,6 +21,7 @@ from .reader.phenocycler import phenocycler
 from .reader.hyperion import hyperion
 from .reader.utils import ome_tif
 from .reader.wsi import wsi, wsi_autoscale
+from .reader.aics import aicsimageio
 from .report import write_report
 
 from ..utils.data import blobs, uniform